--- Carts-Feature/serenity.properties
@@ -1,1 +1,0 @@
-serenity.project.name=Serenity and Cucumber Quick Start
