--- Carts-Feature/src/test/java/starter/Carts/dummyJsonResponses.java
@@ -1,8 +1,0 @@
-package starter.Carts;
-
-public class dummyJsonResponses {
-    public static final String ID = "id";
-    public static final String USERID = "userId";
-    public static final String PRODUCTS = "products";
-    public static final String QUANTITY = "quantity";
-}
