--- Carts-Feature/src/test/java/starter/CucumberTestSuite.java
@@ -1,14 +1,0 @@
-package starter;
-
-
-import io.cucumber.junit.CucumberOptions;
-import net.serenitybdd.cucumber.CucumberWithSerenity;
-import org.junit.runner.RunWith;
-
-@RunWith(CucumberWithSerenity.class)
-@CucumberOptions(
-        plugin = {"pretty"},
-        tags = "@tugas",
-        features = "classpath:features"
-)
-public class CucumberTestSuite {}
