--- Carts-Feature/src/test/resources/serenity.conf
@@ -1,20 +1,0 @@
-#
-# This section defines environment-specific configuration for different environments.
-# You can define normal Serenity properties, such as webdriver.base.url, or custom ones
-# You can find more details about this feature at https://johnfergusonsmart.com/environment-specific-configuration-in-serenity-bdd/
-#
-
-environments {
-  default {
-    webdriver.base.url = "https://duckduckgo.com"
-  }
-  dev {
-    webdriver.base.url = "https://duckduckgo.com/dev"
-  }
-  staging {
-    webdriver.base.url = "https://duckduckgo.com/staging"
-  }
-  prod {
-    webdriver.base.url = "https://duckduckgo.com/prod"
-  }
-}
