+++ src/test/java/starter/DummyJson/dummyJsonAPI.java
@@ -42,6 +42,20 @@ public class dummyJsonAPI {
     //Scenario DELETE
     public static String DELETE_POST = URL + "/posts/{id}";
 
+    public static String GET_ALL_CARTS = URL + "/carts";
+    public static String GET_SINGLE_CARTS = URL + "/carts/{id}";
+    public static String GET_CART_USER_BY_USER_ID = URL + "/carts/user/{userId}";
+
+    //Scenario POST
+    public static String POST_CREATE_CART = URL + "/carts/add";
+
+    //Scenario PUT
+    public static String PUT_UPDATE_CART = URL + "/carts/{id}";
+
+    //Scenario DELETE
+    public static String DELETE_CART = URL + "/carts/{id}";
+
+
 
     public static String GET_ALL_PRODUCTS = URL + "/products";
     public static String GET_SINGLE_PRODUCTS = URL + "/products/{id}";
@@ -100,6 +114,46 @@ public class dummyJsonAPI {
     //Scenario Delete Post
     @Step("Delete Post")
     public void setDeletePost(int id){
+        SerenityRest.given().pathParam("id", id);
+    }
+
+    @Step("Get All Carts")
+    public void setGetAllCarts(){
+        SerenityRest.given();
+    }
+
+    //Scenario Get Single Cart
+    @Step("Get Single Cart")
+    public void setGetSingleCarts(int id){
+        SerenityRest.given().pathParam("id", id);
+    }
+
+    //Scenario Cart User By User ID
+    @Step("Get Cart User By User ID")
+    public void setGetCartUserByUserId(int userId){
+        SerenityRest.given().pathParam("userId", userId);
+    }
+
+    //Scenario Post Create Cart
+    @Step("Post Create Cart")
+    public void setPostCreateCart(File json){
+        SerenityRest.given()
+                .contentType(ContentType.JSON)
+                .body(json);
+    }
+
+    //Scenario Put Update Cart
+    @Step("Put Update Cart")
+    public void setPutUpdateCart(int id, File json){
+        SerenityRest.given()
+                .pathParam("id", id)
+                .contentType(ContentType.JSON)
+                .body(json);
+    }
+
+    //Scenario Delete Cart
+    @Step("Delete Cart")
+    public void setDeleteCart(int id){
         SerenityRest.given().pathParam("id", id);
     }
