+++ src/test/resources/JSON/Requestbody/PutUpdateCart.json
@@ -1,0 +1,8 @@
+{
+  "products": [
+    {
+      "id": 4,
+      "quantity": 80
+    }
+  ]
+}
